@@ -339,7 +339,7 @@ class UsersController < OrganizationAwareController
     else
       @user.active = true
     end
-    @user.save(:validate => :false)
+    @user.save(:validate => false)
     respond_to do |format|
       notify_user(:notice, "User #{@user} has been deactivated.")
       format.html { redirect_to user_url(@user) }
