@@ -130,11 +130,11 @@ class Asset < ActiveRecord::Base
   validates     :asset_type_id,       :presence => true
   validates     :asset_subtype_id,    :presence => true
   validates     :created_by_id,       :presence => true
-  validates     :manufacture_year,    :presence => true, :numericality => {:only_integer => :true, :greater_than_or_equal_to => 1900}
-  validates     :expected_useful_life, :numericality => {:only_integer => :true, :greater_than_or_equal_to => 0}, :presence => true
+  validates     :manufacture_year,    :presence => true, :numericality => {:only_integer => true, :greater_than_or_equal_to => 1900}
+  validates     :expected_useful_life, :numericality => {:only_integer => true, :greater_than_or_equal_to => 0}, :presence => true
   validates_inclusion_of :purchased_new, :in => [true, false]
-  validates     :purchase_cost,       :presence => :true, :numericality => {:only_integer => :true, :greater_than_or_equal_to => 0}
-  validates     :purchase_date,       :presence => :true
+  validates     :purchase_cost,       :presence => true, :numericality => {:only_integer => true, :greater_than_or_equal_to => 0}
+  validates     :purchase_date,       :presence => true
   validates     :serial_number,       uniqueness: {scope: :organization, message: "must be unique within an organization"}, allow_nil: true, allow_blank: true
 
   #-----------------------------------------------------------------------------
