@@ -7,7 +7,7 @@ class Component < Asset
   # Associations common to all equipment assets
   #------------------------------------------------------------------------------
 
-  validates :quantity,        :presence => :true, :numericality => {:only_integer => :true, :greater_than_or_equal_to => 1}
+  validates :quantity,        :presence => true, :numericality => {:only_integer => true, :greater_than_or_equal_to => 1}
   validates :quantity_units,  :presence => true
 
   #------------------------------------------------------------------------------
