@@ -22,7 +22,7 @@ class ConditionUpdateEvent < AssetEvent
       :greater_than_or_equal_to => ConditionType.minimum(:rating),
       :less_than_or_equal_to    => ConditionType.maximum(:rating)
       },
-      :allow_nil    => :true
+      :allow_nil    => true
   # validates :comments, :length => {:maximum => ???} # There is no limit in the Database/Schema
 
   before_validation do
