@@ -4,7 +4,9 @@ require 'cancan'
 require 'unitwise'
 require 'chronic'
 require 'breadcrumbs_on_rails'
-require 'state_machine'
+require 'state_machines'
+require 'state_machines-activemodel'
+require 'state_machines-activerecord'
 require 'kaminari'
 require 'delayed_job'
 require 'delayed_job_active_record'
